refactor to add wish list button to correct component
