+++ client/src/components/App.jsx
@@ -4,7 +4,6 @@ import styled from 'styled-components';
 import Carousel from './Carousel';
 import LeftArrow from './LeftArrow';
 import RightArrow from './RightArrow';
-import WishListButton from './WishListButton';
 
 const Title = styled.h2`
   font-family: Roboto, sans-serif;
@@ -128,7 +127,6 @@ class App extends React.Component {
           {this.renderRightArrow()}
         </CarouselContainer>
         <Title>Wish List</Title>
-        <WishListButton />
       </div>
     );
   }
