+++ client/src/components/product-components/info/Info.jsx
@@ -2,11 +2,19 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import renderImpacts from './impacts/renderImpacts';
-import WishListButton from '../../WishListButton';
+import WishListButton from '../colors/WishListButton';
 
 const Wrapper = styled.div`
   padding-top: 10px;
   padding-left: 20px;
+`;
+
+const Name = styled.div`
+  font-family: Roboto, sans-serif;
+  font-weight: 500;
+  font-size: 14px;
+  width: 280px;
+  height: 32px;
 `;
 
 const BottomWrapper = styled.div`
@@ -21,19 +29,6 @@ const BottomLeftWrapper = styled.div`
   height: 60px;
 `;
 
-const ButtonWrapper = styled.div`
-  padding-top: 15px;
-  padding-right: 10px;
-`;
-
-const Name = styled.div`
-  font-family: Roboto, sans-serif;
-  font-weight: 500;
-  font-size: 14px;
-  width: 280px;
-  height: 32px;
-`;
-
 const Price = styled.div`
   font-family: Roboto, sans-serif;
   font-weight: 400;
@@ -46,6 +41,11 @@ const Colors = styled.div`
   color: grey;
   font-weight: 100;
   font-size: 12px;
+`;
+
+const ButtonWrapper = styled.div`
+  padding-top: 15px;
+  padding-right: 10px;
 `;
 
 const Info = (props) => {
